Last Commit

+++ src/main/java/com/keyin/member/MemberArrayList.java
@@ -1,5 +1,6 @@
 package com.keyin.member;
 import java.io.*;
+import java.nio.file.StandardOpenOption;
 import java.util.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -46,24 +47,30 @@ public class MemberArrayList {
                     System.out.print("Phone " + phone.get(i) );
                     System.out.println("");
                 }
-               Path output = Paths.get("/Users/keyinstudent/IdeaProjects/sprint1/src/main/java/com/keyin/member/ArrayOutput.txt");
-                //Path output = Paths.get("/Users/MXZ User/Desktop/SD_Architecture_&_Design/SDAT_Week_6/sprint1/src/main/java/com/keyin/member/ArrayOutput.txt");
+//                Path output = Paths.get("/Users/MXZ User/Desktop/MemberArrayOutput.txt");
+                String output_path = "/Users/MXZ User/Desktop/SD_Architecture_&_Design/SDAT_Week_6/sprint1/src/main/java/com/keyin/member/MemberArrayOutput.txt";
+                Path output = Paths.get(output_path);
                 try {
                     String newline = System.getProperty("line.separator");
                     StringBuilder str = new StringBuilder();
-                    str.append(firstName ).append(newline );
-                    str.append(lastName ).append(newline  );
-                    str.append(memberId ).append(newline );
-                    str.append(email ).append(newline );
-                    str.append(phone );
-                    Files.write(output, Collections.singleton(str.toString()));
-                    System.out.println(output.toFile().getAbsolutePath());
+                    str.append("First Name: " + firstName ).append(newline );
+                    str.append("Last Name: " + lastName ).append(newline  );
+                    str.append("Member Id: " + memberId ).append(newline );
+                    str.append("Email: " + email ).append(newline );
+                    str.append("Phone: " + phone );
+                    if(new File(output_path).isFile()) {
+                        Files.write(output, Collections.singleton(str.toString()), StandardOpenOption.APPEND);
+                    }
+                    else {
+                        Files.write(output, Collections.singleton(str.toString()));
+                    }
+
+                    System.out.format(output.toFile().getAbsolutePath());
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
 
             }
-            }
         }
     }
-
+}
